Error with p=0
